requirements.txt
gspread
pandas
psycopg2-binary
python-dotenv
requests
slack-bolt
